refactor: GoogleMapDrawer 구현
