+++ android/app/src/main/java/io/coursepick/coursepick/presentation/map/google/GoogleMapManager.kt
@@ -1,28 +1,14 @@
 package io.coursepick.coursepick.presentation.map.google
 
-import android.animation.ValueAnimator
 import android.content.Context
-import android.graphics.Bitmap
-import android.graphics.BitmapFactory
 import android.graphics.Point
-import androidx.annotation.DrawableRes
-import androidx.core.graphics.scale
 import com.google.android.gms.maps.CameraUpdateFactory
 import com.google.android.gms.maps.GoogleMap
-import com.google.android.gms.maps.model.BitmapDescriptorFactory
 import com.google.android.gms.maps.model.CameraPosition
-import com.google.android.gms.maps.model.Circle
-import com.google.android.gms.maps.model.CircleOptions
 import com.google.android.gms.maps.model.LatLng
 import com.google.android.gms.maps.model.LatLngBounds
 import com.google.android.gms.maps.model.MapStyleOptions
-import com.google.android.gms.maps.model.Marker
-import com.google.android.gms.maps.model.MarkerOptions
 import com.google.android.gms.maps.model.Polyline
-import com.google.android.gms.maps.model.PolylineOptions
-import com.google.android.gms.maps.model.StrokeStyle
-import com.google.android.gms.maps.model.StyleSpan
-import com.google.android.gms.maps.model.TextureStyle
 import io.coursepick.coursepick.R
 import io.coursepick.coursepick.domain.course.Coordinate
 import io.coursepick.coursepick.domain.course.Scope
@@ -49,17 +35,7 @@ class GoogleMapManager(
             return Scope(distance)
         }
 
-    private val fineUserLocationImage by lazy {
-        scaleDrawable(R.drawable.image_current_location, 0.5F)
-    }
-    private val searchCoordinateImage by lazy {
-        scaleDrawable(R.drawable.image_search_location, 0.5F)
-    }
-
-    private val polylines = mutableListOf<Polyline>()
-    private var searchCoordinateMarker: Marker? = null
-    private var fineUserLocationMarker: Marker? = null
-    private var coarseUserLocationCircle: Circle? = null
+    private val drawer = GoogleMapDrawer(map, context)
 
     override fun startMap(onMapReady: () -> Unit) {
         map.setMapStyle(MapStyleOptions.loadRawResourceStyle(context, R.raw.google_map_style))
@@ -77,159 +53,35 @@ class GoogleMapManager(
         onMapReady()
     }
 
-    private fun drawCourse(course: CourseItem) {
-        val baseOptions =
-            PolylineOptions()
-                .add(*course.coordinates.map(Coordinate::toLatLng).toTypedArray())
-
-        if (course.selected) {
-            val selectedCourseOptions: PolylineOptions =
-                baseOptions
-                    .color(context.getColor(R.color.course_selected))
-                    .width(context.resources.getDimension(R.dimen.selected_course_width))
-                    .zIndex(SELECTED_COURSE_Z_INDEX)
-            map
-                .addPolyline(selectedCourseOptions)
-                .also(polylines::add)
-
-            val selectedCourseOverlayOptions: PolylineOptions =
-                baseOptions
-                    .width(context.resources.getDimension(R.dimen.selected_course_width) * 1.5F)
-                    .addSpan(
-                        StyleSpan(
-                            StrokeStyle
-                                .transparentColorBuilder()
-                                .stamp(
-                                    TextureStyle
-                                        .newBuilder(BitmapDescriptorFactory.fromResource(R.drawable.image_arrow))
-                                        .build(),
-                                ).build(),
-                        ),
-                    ).zIndex(SELECTED_COURSE_Z_INDEX)
-                    .clickable(true)
-            map
-                .addPolyline(selectedCourseOverlayOptions)
-                .apply { tag = course }
-                .also(polylines::add)
-        } else {
-            val unselectedCourseOptions: PolylineOptions =
-                baseOptions
-                    .color(context.getColor(R.color.course_unselected))
-                    .width(context.resources.getDimension(R.dimen.unselected_course_width))
-                    .zIndex(UNSELECTED_COURSE_Z_INDEX)
-                    .clickable(true)
-            map
-                .addPolyline(unselectedCourseOptions)
-                .apply { tag = course }
-                .also(polylines::add)
-        }
-    }
-
     override fun draw(course: CourseItem) {
-        removeAllRouteLines()
-        drawCourse(course)
+        drawer.drawCourse(course)
     }
 
     override fun draw(courses: List<CourseItem>) {
-        removeAllRouteLines()
-        courses.forEach(::drawCourse)
+        courses.forEach(drawer::drawCourse)
     }
 
     override fun drawRouteToCourse(
         route: List<Coordinate>,
         course: CourseItem,
     ) {
-        removeAllRouteLines()
-        map
-            .addPolyline(
-                PolylineOptions()
-                    .add(*route.map(Coordinate::toLatLng).toTypedArray())
-                    .width(context.resources.getDimension(R.dimen.course_route_width))
-                    .color(context.getColor(R.color.course_route)),
-            ).also(polylines::add)
-        drawCourse(course)
+        drawer.drawRouteToCourse(route, course)
     }
 
     override fun removeAllRouteLines() {
-        polylines.forEach(Polyline::remove)
-        polylines.clear()
+        drawer.removeAllRouteLines()
     }
 
     override fun drawSearchCoordinate(coordinate: Coordinate) {
-        searchCoordinateMarker?.let { marker: Marker ->
-            marker.position = coordinate.toLatLng()
-        } ?: run {
-            searchCoordinateMarker =
-                map.addMarker(
-                    MarkerOptions()
-                        .icon(BitmapDescriptorFactory.fromBitmap(searchCoordinateImage))
-                        .position(map.cameraPosition.target)
-                        .anchor(0.5F, 1F),
-                )
-        }
+        drawer.drawSearchCoordinate(coordinate)
     }
 
     override fun drawUserLocation(location: Location) {
-        when (location) {
-            is Location.Fine -> drawFineUserLocation(location)
-            is Location.Coarse -> drawCoarseUserLocation(location)
-        }
-    }
-
-    private fun drawFineUserLocation(location: Location.Fine) {
-        hideCoarseUserLocation()
-
-        fineUserLocationMarker?.let { marker: Marker ->
-            animateLatLng(
-                start = marker.position,
-                end = location.coordinate.toLatLng(),
-                duration = MOVE_ANIMATION_DURATION_MS,
-            ) { latLng: LatLng -> marker.position = latLng }
-        } ?: run {
-            fineUserLocationMarker =
-                map.addMarker(
-                    MarkerOptions()
-                        .icon(BitmapDescriptorFactory.fromBitmap(fineUserLocationImage))
-                        .position(location.coordinate.toLatLng())
-                        .anchor(0.5F, 0.5F),
-                )
-        }
-    }
-
-    private fun drawCoarseUserLocation(location: Location.Coarse) {
-        hideFineUserLocation()
-
-        coarseUserLocationCircle?.let { marker: Circle ->
-            animateLatLng(
-                start = marker.center,
-                end = location.coordinate.toLatLng(),
-                duration = MOVE_ANIMATION_DURATION_MS,
-            ) { latLng: LatLng -> marker.center = latLng }
-        } ?: run {
-            coarseUserLocationCircle =
-                map.addCircle(
-                    CircleOptions()
-                        .center(location.coordinate.toLatLng())
-                        .radius(location.accuracy.meter.value)
-                        .fillColor(context.getColor(R.color.coarse_location_area))
-                        .strokeWidth(0F),
-                )
-        }
+        drawer.drawUserLocation(location)
     }
 
     override fun hideUserLocation() {
-        hideFineUserLocation()
-        hideCoarseUserLocation()
-    }
-
-    private fun hideFineUserLocation() {
-        fineUserLocationMarker?.remove()
-        fineUserLocationMarker = null
-    }
-
-    private fun hideCoarseUserLocation() {
-        coarseUserLocationCircle?.remove()
-        coarseUserLocationCircle = null
+        drawer.hideUserLocation()
     }
 
     override fun fitTo(coordinates: List<Coordinate>) {
@@ -237,7 +89,7 @@ class GoogleMapManager(
             CameraUpdateFactory.newLatLngBounds(
                 LatLngBounds
                     .builder()
-                    .apply { coordinates.map(Coordinate::toLatLng).forEach(::include) }
+                    .apply { coordinates.forEach { coordinate: Coordinate -> this.include(coordinate.toLatLng()) } }
                     .build(),
                 context.resources.getDimensionPixelSize(R.dimen.course_route_padding),
             ),
@@ -281,38 +133,11 @@ class GoogleMapManager(
         map.setPadding(left, top, right, bottom)
     }
 
-    private fun animateLatLng(
-        start: LatLng,
-        end: LatLng,
-        duration: Long,
-        onChange: (latLng: LatLng) -> Unit,
-    ) {
-        val valueAnimator = ValueAnimator.ofFloat(0F, 1F).setDuration(duration)
-        valueAnimator.addUpdateListener { animator: ValueAnimator ->
-            val latitude =
-                (end.latitude - start.latitude) * animator.animatedFraction + start.latitude
-            val longitude =
-                (end.longitude - start.longitude) * animator.animatedFraction + start.longitude
-            onChange(LatLng(latitude, longitude))
-        }
-        valueAnimator.start()
-    }
-
-    private fun scaleDrawable(
-        @DrawableRes id: Int,
-        scale: Float,
-    ): Bitmap {
-        val original: Bitmap = BitmapFactory.decodeResource(context.resources, id)
-        return original.scale((original.width * scale).toInt(), (original.height * scale).toInt())
-    }
-
     companion object {
         private const val MOVE_ANIMATION_DURATION_MS = 750L
         private const val DEFAULT_LATITUDE = 37.5100226
         private const val DEFAULT_LONGITUDE = 127.1026170
         private val DEFAULT_LATLNG = LatLng(DEFAULT_LATITUDE, DEFAULT_LONGITUDE)
         private const val DEFAULT_ZOOM_LEVEL = 15F
-        private const val SELECTED_COURSE_Z_INDEX = 1F
-        private const val UNSELECTED_COURSE_Z_INDEX = 0F
     }
 }
